imported JSDOM the proper way
