+++ frontend/package.json
@@ -15,6 +15,7 @@
     "@testing-library/jest-dom": "^6.6.3",
     "@testing-library/svelte": "^5.2.7",
     "@tsconfig/svelte": "^5.0.4",
+    "@types/jsdom": "^21.1.7",
     "@vitest/ui": "^3.1.2",
     "jsdom": "^26.1.0",
     "svelte": "^5.23.1",
